Restructure tests to support mocking with HippoMocks.

+++ include/microlib.h
@@ -55,7 +55,7 @@ static const int SUCCESS = 0;
   void putc(char c, void *stream);
   extern int puts(const char * s);
   size_t strlen(const char *s);
-  size_t memcmp(const void *s1, const void *s2, size_t n);
+  int memcmp(const void *s1, const void *s2, size_t n);
   size_t strnlen(const char *s, size_t max);
   void * memchr(const void *s, int c, size_t n);
   void * memset(void *b, int c, size_t len);
